backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
pypdf
ebooklib
beautifulsoup4
